Refactor HasDescription
