+++ betty/plugins/entity/event.py
@@ -15,7 +15,6 @@ from betty.entity.association import (
 )
 from betty.entity.has_citations import HasCitations
 from betty.entity.has_date import HasDate
-from betty.entity.has_description import HasDescription
 from betty.entity.has_file_references import HasFileReferences
 from betty.entity.has_links import HasLinks
 from betty.entity.has_notes import HasNotes
@@ -32,6 +31,7 @@ from betty.plugins.entity.presence import Presence
 from betty.plugins.event_type.unknown import Unknown as UnknownEventType
 from betty.plugins.role.subject import Subject
 from betty.privacy import HasPrivacy, Privacy
+from betty.properties.description import HasDescription
 from betty.properties.localizable import LocalizableProperty
 from betty.property import Optional
