--- betty/test_utils/ancestry/description.py
@@ -1,11 +1,0 @@
-"""
-Test utilities for :py:mod:`betty.entity.has_description`.
-"""
-
-from betty.entity.has_description import HasDescription
-
-
-class DummyHasDescription(HasDescription):
-    """
-    A dummy :py:class:`betty.entity.has_description.HasDescription` implementation.
-    """
